pagination works, most rspec tests still pass.
We'll need to update our DB schemas next, because:

If you pry into the `trips.each` block, inside of views/trips/index.erb,
you'll see that the trip data is returning nils for a view values.
Specifically:

- start_station_name
- end_station_name

I'm not positive yet of the fix here, but we'll get it tomorrow.