Remove extraneous model specs
This commit removes specs which were intended to verify that Station attributes have the proper input types. However, we learned that Postgres and/or ActiveRecord converts and saves values in the database according to the type that is specified in the schema, so trying to test Stations by inputting improper data types as attributes will not work. Those improper data types are converted into the proper type in order to save the record. With this knowledge, we have decided that tests for invalid Station attribute types are not relevant.